Rename manager to repository
